Build a lazy seq of survivors with partial results
Build a lazy seq of partial results of the mutation analysis instead of
a single persistent map with a conclusive result.

The last element of the returning lazy seq should be identical to a map
with a conclusive result. Previous elements of the seq could be used to
report progress of the analysis.

Since the collection of results is lazy, a mutation analysis session can be
interrupted upon running into the first survivor by simply not evaluating
the rest of the seq.